requirements.txt
python=3.9.12
pytorch=1.11.0
transformers=4.18.0
numpy=1.22.3
datasets=2.1.0
pandas=1.4.2
spacy=3.2.4



